+++ include/leo/custom_antenna.h
@@ -53,9 +53,9 @@ public:
    * \return a boost::shared_ptr to the constructed tracker object.
    */
   static generic_antenna::generic_antenna_sptr
-  make(uint8_t type, float frequency, int polarization,
-       float pointing_error, float gain, float beamwidth,
-       float rolloff_gain);
+  make(uint8_t type, double frequency, int polarization,
+       double pointing_error, double gain, double beamwidth,
+       double rolloff_gain);
 
   custom_antenna() {};
