+++ include/leo/link_margin.h
@@ -39,10 +39,10 @@ public:
   static link_margin::link_margin_sptr
   make();
 
-  float
-  calc_link_margin(float total_loss_db, float satellite_antenna_gain,
-                   float tracker_antenna_gain,
-                   float transmission_power_dbw, float noise_floor);
+  double
+  calc_link_margin(double total_loss_db, double satellite_antenna_gain,
+                   double tracker_antenna_gain,
+                   double transmission_power_dbw, double noise_floor);
 
 };
 } // namespace leo
