+++ include/leo/monopole_antenna.h
@@ -49,7 +49,7 @@ public:
    * \return a boost::shared_ptr to the constructed tracker object.
    */
   static generic_antenna::generic_antenna_sptr
-  make(uint8_t type, float frequency, int polarization, float pointing_error);
+  make(uint8_t type, double frequency, int polarization, double pointing_error);
 
   monopole_antenna() {};
